Multivariate Logistic Random Variables

Computes random vectors of the multivariate symmetric logistic distribution for arbitrary correlation matrices using the asymptotic Kolmogorov distribution – see references.

Usage

rmvlogis(n, Q = NULL, delta = rep(0, d), BIG = 500)

Arguments

n

number of observations

Q

semi-positive definite

delta

location vector.

BIG

the number of exponential to add for asymptotic Kolomogrov

References

Scale Mixtures of Normal Distributions Author(s): D. F. Andrews and C. L. Mallows Source: Journal of the Royal Statistical Society. Series B (Methodological), Vol. 36, No. 1 (1974), pp. 99-102 Published by: Wiley for the Royal Statistical Society Stable URL: http://www.jstor.org/stable/2984774

Examples

rmvlogis(10, Q=diag(5))
#>             [,1]       [,2]       [,3]       [,4]        [,5]
#>  [1,]  2.5890076 -2.2490857 -0.7472117  3.0222297 -2.07197993
#>  [2,] -1.4574477 -1.5208663  0.3222531  0.4034215  2.37985890
#>  [3,] -0.3356744  4.1591546 -1.3953634  2.1618945  1.10518635
#>  [4,]  2.6008694  1.0353735 -1.5661405  0.9971147  0.54545779
#>  [5,] -2.9056077 -2.8728200  3.4639660  1.1003938  0.87503101
#>  [6,]  0.8176184 -2.5421209 -0.8326681 -1.2896730  2.40667587
#>  [7,] -1.9007258  0.6085258 -0.7593202  0.1130552  0.65678665
#>  [8,]  0.9265653  0.8998055  0.9115432 -1.1959522 -0.09550614
#>  [9,]  1.7208763 -1.4449115  1.2963924 -2.5018370 -3.01443119
#> [10,] -0.1689555  0.1927892  0.2790179 -0.6932005  1.04779718

if (FALSE) { # \dontrun{
QMAT <- matrix(c(1,0,0,1),nrow=2)
draw_NNMD  <- NonNorMvtDist::rmvlogis(2e3, parm1=rep(0,2), parm2=rep(1,2))
draw_mvpd  <-          mvpd::rmvlogis(2e3,     Q=QMAT)

mean(draw_NNMD[,1]   < -1 & draw_NNMD[,2]   < 3)
mean(draw_mvpd[,1] < -1 & draw_mvpd[,2] < 3)

plogis(-1)
mean(draw_NNMD[,1] < -1)
mean(draw_mvpd[,1] < -1)

plogis(3)
mean(draw_NNMD[,2] < 3)
mean(draw_mvpd[,2] < 3)
 
rangex <- range(c(draw_mvpd[,1],draw_NNMD[,1]))
rangey <- range(c(draw_mvpd[,2],draw_NNMD[,2]))

par(mfrow=c(3,2), pty="s", mai=c(.5,.1,.1,.1))
plot(draw_NNMD, xlim=rangex, ylim=rangey); abline(h=0,v=0)
plot(draw_mvpd   , xlim=rangex, ylim=rangey); abline(h=0,v=0)

hist(draw_NNMD[,1]  , breaks = 100,  xlim=rangex, probability=TRUE, ylim=c(0,.40))
curve(dlogis(x), add=TRUE, col="blue",lwd=2)
hist(draw_mvpd[,1], breaks = 100,  xlim=rangex, probability=TRUE, ylim=c(0,.40))
curve(dlogis(x), add=TRUE, col="blue",lwd=2)
hist(draw_NNMD[,2]  , breaks = 100,  xlim=rangex, probability=TRUE, ylim=c(0,.40))
curve(dlogis(x), add=TRUE, col="blue",lwd=2)
hist(draw_mvpd[,2], breaks = 100,  xlim=rangex, probability=TRUE, ylim=c(0,.40))
curve(dlogis(x), add=TRUE, col="blue",lwd=2)
} # }