Multivariate Subgaussian Stable Random Variates

Computes random vectors of the multivariate subgaussian stable distribution for arbitrary alpha, shape matrices, and location vectors. See Nolan (2013).

Usage

rmvss(
  n,
  alpha = 1,
  Q = NULL,
  delta = rep(0, d),
  which.stable = c("libstable4u", "stabledist")[1]
)

Arguments

n

number of observations

alpha

default to 1 (Cauchy). Must be 0<alpha<2

Q

Shape matrix. See Nolan (2013).

delta

location vector.

which.stable

defaults to "libstable4u", other option is "stabledist". Indicates which package should provide the univariate stable distribution in this production distribution form of a univariate stable and multivariate normal.

Value

Returns the n by d matrix containing multivariate subgaussian stable random variates where d=nrow(Q).

References

Nolan JP (2013), Multivariate elliptically contoured stable distributions: theory and estimation. Comput Stat (2013) 28:2067–2089 DOI 10.1007/s00180-013-0396-7

Examples

## generate 10 random variates of a bivariate mvss
rmvss(n=10, alpha=1.71, Q=matrix(c(10,7.5,7.5,10),2))
#>              [,1]       [,2]
#>  [1,]  -3.1547271  -2.103857
#>  [2,]  -3.7144064  -5.604734
#>  [3,]   5.2082938   2.551250
#>  [4,]  -0.2412486   3.810356
#>  [5,]   1.7277478   1.988887
#>  [6,] -23.5193885 -40.459124
#>  [7,]   4.8197852   7.089041
#>  [8,]   0.2690873   1.446094
#>  [9,]  -2.1926593  -7.589429
#> [10,]   7.2049296   5.295507

## generate 10 random variates of a trivariate mvss
Q <- matrix(c(10,7.5,7.5,7.5,10,7.5,7.5,7.5,10),3)
rmvss(n=10, alpha=1.71, Q=Q)
#>              [,1]         [,2]       [,3]
#>  [1,]  -1.1757654  -2.43884743  4.5784581
#>  [2,]   3.2302965  -0.29315919 -1.3906679
#>  [3,]   1.5894881  -1.22109764 -0.3356316
#>  [4,]   0.9949992  -0.06168804 -4.0382708
#>  [5,]  -3.6108905 -11.65373754 -4.5608165
#>  [6,]  -1.8763509  -2.69880449 -4.4783599
#>  [7,]   3.6767780   7.68304748  3.1228319
#>  [8,] -12.3291686  -5.18658197 -7.2320985
#>  [9,]  11.1562491   5.19857597  2.3980446
#> [10,]   3.2788009   4.33402699  1.8478028