Multivariate Logistic Random Variables

Computes random vectors of the multivariate symmetric logistic distribution for arbitrary correlation matrices using the asymptotic Kolmogorov distribution – see references.

Usage

rmvlogis(n, Q = NULL, delta = rep(0, d), BIG = 500)

Arguments

n

number of observations

Q

semi-positive definite

delta

location vector.

BIG

the number of exponential to add for asymptotic Kolomogrov

References

Scale Mixtures of Normal Distributions Author(s): D. F. Andrews and C. L. Mallows Source: Journal of the Royal Statistical Society. Series B (Methodological), Vol. 36, No. 1 (1974), pp. 99-102 Published by: Wiley for the Royal Statistical Society Stable URL: http://www.jstor.org/stable/2984774

Examples

rmvlogis(10, Q=diag(5))
#>             [,1]       [,2]       [,3]       [,4]        [,5]
#>  [1,] -1.4554436 -0.5413883 -0.8667427 -0.5031270  1.36947506
#>  [2,] -0.8491013 -0.8416934  1.8264470  0.6601506  0.98888253
#>  [3,] -2.8792461  1.0411447  1.8077202  1.6215116 -0.07824789
#>  [4,]  0.8069226 -0.3484485 -0.4003437  1.2489126 -0.60357591
#>  [5,] -1.5304412 -1.9614416  0.6800465 -2.5021227  2.64421762
#>  [6,]  0.0658620  0.1420186  1.4567006  0.3147913  0.60365273
#>  [7,]  0.1771488 -0.2844730 -1.3069375 -1.0245978  0.42505409
#>  [8,] -0.9856645 -0.7346512 -0.3280713  0.3037126  2.69082351
#>  [9,]  1.9138452 -0.8799090 -2.9995169 -2.9453641  2.68784938
#> [10,]  2.0868364 -0.5954600  0.6351749 -0.6499376 -1.92336192

if (FALSE) { # \dontrun{
QMAT <- matrix(c(1,0,0,1),nrow=2)
draw_NNMD  <- NonNorMvtDist::rmvlogis(2e3, parm1=rep(0,2), parm2=rep(1,2))
draw_mvpd  <-          mvpd::rmvlogis(2e3,     Q=QMAT)

mean(draw_NNMD[,1]   < -1 & draw_NNMD[,2]   < 3)
mean(draw_mvpd[,1] < -1 & draw_mvpd[,2] < 3)

plogis(-1)
mean(draw_NNMD[,1] < -1)
mean(draw_mvpd[,1] < -1)

plogis(3)
mean(draw_NNMD[,2] < 3)
mean(draw_mvpd[,2] < 3)
 
rangex <- range(c(draw_mvpd[,1],draw_NNMD[,1]))
rangey <- range(c(draw_mvpd[,2],draw_NNMD[,2]))

par(mfrow=c(3,2), pty="s", mai=c(.5,.1,.1,.1))
plot(draw_NNMD, xlim=rangex, ylim=rangey); abline(h=0,v=0)
plot(draw_mvpd   , xlim=rangex, ylim=rangey); abline(h=0,v=0)

hist(draw_NNMD[,1]  , breaks = 100,  xlim=rangex, probability=TRUE, ylim=c(0,.40))
curve(dlogis(x), add=TRUE, col="blue",lwd=2)
hist(draw_mvpd[,1], breaks = 100,  xlim=rangex, probability=TRUE, ylim=c(0,.40))
curve(dlogis(x), add=TRUE, col="blue",lwd=2)
hist(draw_NNMD[,2]  , breaks = 100,  xlim=rangex, probability=TRUE, ylim=c(0,.40))
curve(dlogis(x), add=TRUE, col="blue",lwd=2)
hist(draw_mvpd[,2], breaks = 100,  xlim=rangex, probability=TRUE, ylim=c(0,.40))
curve(dlogis(x), add=TRUE, col="blue",lwd=2)
} # }